// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const EstadoCita = {
  PROGRAMADA: 'PROGRAMADA',
  COMPLETADA: 'COMPLETADA',
  CANCELADA: 'CANCELADA'
} as const

export type EstadoCita = (typeof EstadoCita)[keyof typeof EstadoCita]


export const Rol = {
  RECEPCIONISTA: 'RECEPCIONISTA',
  MEDICO: 'MEDICO',
  GERENCIA: 'GERENCIA'
} as const

export type Rol = (typeof Rol)[keyof typeof Rol]